Fixed edge case when reshaping hough lines
HoughLines produces arrays on the shape (num_lines, 1, 2)
which we want to reshape to (num_lines, 2). Previously
np.squeeze was used here to remove the extra dimension.
However, if the array has size (1, 1, 2) the squeeze
operation would remove both first dimensions and give the
shape (2, ) only, which then caused following code to
crash. This is now replaces with a reshape instead.